% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinants.R
\name{adjoint}
\alias{adjoint}
\title{Calculate the Adjoint of a matrix}
\usage{
adjoint(A)
}
\arguments{
\item{A}{a square matrix}
}
\value{
a matrix of the same size as \code{A}
}
\description{
This function calculates the adjoint of a square matrix, defined as the transposed
matrix of cofactors of all elements.
}
\examples{
A <- J(3, 3) + 2*diag(3)
adjoint(A)
}
\seealso{
Other determinants: 
\code{\link{Det}()},
\code{\link{cofactor}()},
\code{\link{minor}()},
\code{\link{rowCofactors}()},
\code{\link{rowMinors}()}
}
\author{
Michael Friendly
}
\concept{determinants}
