% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrows3d.R
\name{cone3d}
\alias{cone3d}
\title{Title Draw a 3d cone}
\usage{
cone3d(base, tip, radius = 10, col = "grey", scale = NULL, ...)
}
\arguments{
\item{base}{coordinates of base of the cone}

\item{tip}{coordinates of tip of the cone}

\item{radius}{radius of the base}

\item{col}{color}

\item{scale}{scale factor for base and tip}

\item{...}{rgl arguments passed down}
}
\value{
returns the integer object ID of the shape that was added to the scene
}
\description{
Title Draw a 3d cone
}
\examples{
# none yet
}
\author{
January Weiner, from pca3d
}

