% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian-elimination.R
\name{Det}
\alias{Det}
\title{Determinant of a Square Matrix}
\usage{
Det(X, method = c("elimination", "eigenvalues"), ...)
}
\arguments{
\item{X}{a square matrix}

\item{method}{one of `"elimination"` (the default) or `"eigenvalues"`}

\item{...}{arguments passed to \code{\link{gaussianElimination}} or \code{\link{Eigen}}}
}
\value{
the determinant of \code{X}
}
\description{
Returns the determinant of a square matrix \code{X},
computed either by Gaussian elimination or as the product of the eigenvalues of the matrix.
If the latter, \code{X} must be symmstric.
}
\examples{
A <- matrix(c(1,2,3,2,5,6,3,6,10), 3, 3) # nonsingular, symmetric
A
Det(A)
B <- matrix(1:9, 3, 3) # a singular matrix
B
Det(B)
C <- matrix(c(1, .5, .5, 1), 2, 2) # square, symmetric, nonsingular
Det(C)
Det(C, method="eigenvalues")
}
\author{
John Fox
}
\seealso{
\code{\link[base]{det}} for the base R function

\code{\link{gaussianElimination}}, \code{\link{Eigen}}
}

