% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{dot}
\alias{dot}
\title{Dot Product of Latex Vectors}
\usage{
dot(x, y, simplify = TRUE)
}
\arguments{
\item{x}{A numeric or character vector}

\item{y}{A numeric or character vector. The lengths of \code{x} and \code{y} must
be equal}

\item{simplify}{logical; if \code{TRUE}, terms involving the numeric constants 
\code{-1, 0, 1} are simplified}
}
\value{
A character string containing the LaTeX expression
}
\description{
The result of matrix multiplication, \eqn{\mathbf{C} = \mathbf{A} \: \mathbf{B}}
is composed of the vector dot products of each \emph{row} of \eqn{\mathbf{A}} with
each \emph{column} of \eqn{\mathbf{B}},
\deqn{c_{ij} = \mathbf{a}_i^\top \; \mathbf{b}_j 
             = \Sigma_k a_{ik} \cdot b_{kj}}

This function computes this product symbolically in LaTeX notation for
numeric and character vectors, simplifying the result if \code{simplify = TRUE}.
The LaTeX symbol for multiplication (\code{"\\cdot"} by default)
can be changed by changing \code{options(latexMultSymbol)},
e.g, \code{options(latexMultSymbol = "\\times")}.
}
\examples{
num <- -1:2
chr <- letters[1:4]

dot(num, num)
dot(num, chr)
dot(chr, num)
dot(chr, chr)

dot(num, num, simplify = FALSE) 
dot(num, chr, simplify = FALSE) 
dot(chr, num, simplify = FALSE) 
dot(chr, chr, simplify = FALSE)

# change the multpliicatin symbol
opt <- options(latexMultSymbol = "\\\\times")
dot(num, chr)
dot(chr, chr)
}
