% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_square.R
\name{is_square_matrix}
\alias{is_orthogonal_matrix}
\alias{is_square_matrix}
\alias{is_symmetric_matrix}
\title{Test for square matrix}
\usage{
is_square_matrix(x)

is_symmetric_matrix(x)

is_orthogonal_matrix(x)
}
\arguments{
\item{x}{A numeric matrix}

\item{x}{A numeric matrix}

\item{x}{A numeric matrix}
}
\value{
\code{TRUE} if \code{x} is a square matrix, else \code{FALSE}

\code{TRUE} if \code{x} is a square, symmetric matrix, else \code{FALSE}

\code{TRUE} if \code{x} is an orthogonal matrix, else \code{FALSE}
}
\description{
Test for square matrix

Test for square, symmetric matrix

Test for orthogonal matrix
}
\keyword{internal}

