% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_square.R
\name{is_symmetric_matrix}
\alias{is_symmetric_matrix}
\title{Test for square, symmetrix matrix}
\usage{
is_symmetric_matrix(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
\code{TRUE} if \code{x} is a square, symmetrix matrix, else \code{FALSE}
}
\description{
Test for square, symmetrix matrix
}
\keyword{internal}

