% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len.R
\name{len}
\alias{len}
\title{Length of a Vector or Column Lengths of a Matrix}
\usage{
len(X)
}
\arguments{
\item{X}{a numeric vector or matrix}
}
\value{
a scalar or vector containing the length(s)
}
\description{
\code{len} calculates the Euclidean length (also called Euclidean norm) of a vector or the
length of each column of a numeric matrix.
}
\examples{
len(1:3)
len(matrix(1:9, 3, 3))
}
\seealso{
\code{\link[base]{norm}} for more general matrix norms
}

