% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlib.R
\docType{package}
\name{matlib}
\alias{matlib}
\alias{matlib-package}
\title{matlib: Matrix Functions for Teaching and Learning Linear Algebra and Multivariate Statistics.}
\description{
These functions are mainly for tutorial purposes in learning matrix algebra
ideas using R. In some cases, functions are provided for concepts available
elsewhere in R, but where the function call or name is not obvious.  In other
cases, functions are provided to show or demonstrate an algorithm.
}
\details{
These are not meant for production uses. Other methods are more efficient for larger problems.
}

