% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eqn.R
\name{ref}
\alias{ref}
\title{Provide inline reference of equations}
\usage{
ref(label, html_output = knitr::is_html_output())
}
\arguments{
\item{label}{the equation label used within \code{\link{Eqn}} or
defined explicitly in the document}

\item{html_output}{logical; use references for HTML outputs instead
of the LaTeX? Automatically changed for compiled documents
that support \code{knitr}}
}
\description{
Depending on the output type this function will provide the correct
inline wrapper for MathJax or LaTeX equations. This provides more
consistent referencing when switching between HTML and PDF outputs.
}
\examples{

# used inside of Eqn()
Eqn('e = mc^2', label='eq:einstein')

# use within inline block via ```r ref()```
ref('eq:einstein')
ref('eq:einstein', html_output=TRUE)

}
\seealso{
\code{\link{Eqn}}
}
