% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowops.R
\name{rowswap}
\alias{rowswap}
\title{Interchange two rows of a matrix}
\usage{
rowswap(x, from, to)
}
\arguments{
\item{x}{a matrix, often consisting of the coeficient matrix, A, joined with a vector of constants, b.}

\item{from}{source row.}

\item{to}{destination row}
}
\value{
the matrix \code{x}, with rows \code{from} and \code{to} interchanged
}
\description{
This elementary row operation corresponds to interchanging two equations.
}
\seealso{
\code{\link{echelon}}, \code{\link{gaussianElimination}}

Other elementary row operations: \code{\link{rowadd}},
  \code{\link{rowmult}}
}

