% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symb-matrix.R
\name{symb_matrix}
\alias{symb_matrix}
\title{Create a Symbolic Matrix for LaTeX}
\usage{
symb_matrix(
  symbol = "x",
  rows = "n",
  cols = "m",
  brackets = c("p", "b", "B", "v", "V", ""),
  indent = "\\t",
  start = c(1, 1)
)
}
\arguments{
\item{symbol}{A single character, the name of the matrix elements}

\item{rows}{Number of rows, a single character representing rows symbolically, or an integer, generating
that many rows.}

\item{cols}{Number of columns, a single character representing rows symbolically, or an integer, generating
that many columns.}

\item{brackets}{Type of brackets: \code{"p"} uses parentheses \code{"(", ")"};
\code{"b"} uses square brackets \code{"[", "]"};
\code{"B"} uses braces \code{"{", "}"};
\code{"v"} uses vertical bars \code{"|", "|"};
\code{"V"} uses double vertical bars \code{"||", "||"};
\code{""} generates a plain matrix without delimeters}

\item{indent}{characters to indent each line [not yet implemented]}

\item{start}{A vector of length 2 giving the base value for indexing rows and columns, \code{0} or \code{1}.
The default, \code{start=1} generates rows indices from 1 to \code{rows}.
\code{start=0} generates rows indices from 0 to \code{rows-1}.}
}
\description{
Constructs the latex code for a symbolic matrix, like:
\preformatted{
 \\begin{pmatrix}
   x_{11}  & x_{12}  & \\dots  & x_{1m}  \\
   x_{21}  & x_{22}  & \\dots  & x_{2m}  \\
   \\vdots & \\vdots & \\ddots & \\vdots \\
   x_{n1}  & x_{n2}  & \\dots  & x_{nm}
 \\end{pmatrix}
 }

Alternatively, the number of rows and/or columns can be integers, generating a matrix of given size.

As presently designed, the function outputs the LaTeX code to the console, which can then be copied/pasted into a document.
}
\details{
This implementation assumes that the \code{amsmath} package will be available because it uses the shorthands
\code{\\begin{pmatrix}}, ... rather than \code{\\left( ... \\right)}
}
\examples{
symb_matrix("x", rows = "n", cols = "m", brackets = "p")  # default
symb_matrix("\\\\beta", "p", "q")
symb_matrix("\\\\beta", "p", "q", start=0)

# numeric rows/cols
symb_matrix("y", "p", 5)
symb_matrix("y", 4, "q")
symb_matrix("y", 4, 4)
symb_matrix("\\\\beta", 4, 4, start=0)
symb_matrix("\\\\beta", 4, 4, start=c(0,1))
}
\author{
Michael Friendly
}
