% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Trace of a Matrix}
\usage{
tr(X)
}
\arguments{
\item{X}{a numeric matrix}
}
\value{
a numeric value, the sum of \code{diag(X)}
}
\description{
Calculates the trace of a square numeric matrix, i.e., the sum of its diagonal elements
}
\examples{
X <- matrix(1:9, 3, 3)
tr(X)

}

