% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vandermode.R
\name{vandermode}
\alias{vandermode}
\title{Vandermode Matrix}
\usage{
vandermode(x, n)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{a numeric scalar}
}
\value{
a matrix of size \code{length(x)} x \code{n}
}
\description{
The function returns the Vandermode matrix of a numeric vector, \code{x}, whose columns are the vector
raised to the powers \code{0:n}.
}
\examples{
vandermode(1:5, 4)
}
