% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{arc}
\alias{arc}
\title{Draw an arc showing the angle between vectors}
\usage{
arc(p1, p2, p3, d = 0.1, absolute = TRUE, ...)
}
\arguments{
\item{p1}{Starting point of first vector}

\item{p2}{End point of first vector, and also start of second vector}

\item{p3}{End point of second vector}

\item{d}{The distance from \code{p2} along each vector for drawing their corner}

\item{absolute}{logical; if \code{TRUE}, \code{d} is taken as an absolute distance along the vectors; otherwise it
is calculated as a relative distance, i.e., a fraction of the length of the vectors.}

\item{...}{Arguments passed to \code{link[graphics]{lines}} or to \code{link[rgl]{lines3d}}}
}
\value{
none
}
\description{
A utility function for drawing vector diagrams.
}
\details{
In this implementation, the two vectors are specified by three points, \code{p1}, \code{p2}, \code{p3}, meaning
a line from \code{p1} to \code{p2}, and another line from \code{p2} to \code{p3}.
}
\examples{
vec <- rbind(diag(3), c(1,1,1))
rownames(vec) <- c("X", "Y", "Z", "J")
}
\references{
\url{http://math.stackexchange.com/questions/1507248/find-arc-between-two-tips-of-vectors-in-3d}
}

