% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrows3d.R
\name{arrows3d}
\alias{arrows3d}
\title{Draw 3d Arrows}
\usage{
arrows3d(coords, headlength = 0.035, head = "end", scale = NULL,
  radius = NULL, ...)
}
\arguments{
\item{coords}{A 2n x 3 matrix giving the start and end (x,y,z) coordinates of n arrows, in pairs}

\item{headlength}{Length of the arrow heads, in device units}

\item{head}{Position of the arrow head, e.g., "end"}

\item{scale}{scale factor for base and tip of arrow head}

\item{radius}{radius of the base of the arrow head}

\item{...}{rgl arguments passed down to \code{segments3d} and \code{cone3d}}
}
\value{
none
}
\description{
Draws nice 3d arrows with \code{cone3d}s at their tips
}
\examples{
# none yet
}
\author{
January Weiner, from pca3d
}

