% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrows3d.R
\name{cone3d}
\alias{cone3d}
\title{Draw a 3D cone}
\usage{
cone3d(base, tip, radius = 10, col = "grey", scale = NULL, ...)
}
\arguments{
\item{base}{coordinates of base of the cone}

\item{tip}{coordinates of tip of the cone}

\item{radius}{radius of the base}

\item{col}{color}

\item{scale}{scale factor for base and tip}

\item{...}{rgl arguments passed down; see \code{\link[rgl]{rgl.material}}}
}
\value{
returns the integer object ID of the shape that was added to the scene
}
\description{
Draws a cone in 3D from a \code{base} point to a \code{tip} point, with a given \code{radius} at the base.
This is used to draw nice arrow heads in \code{\link{arrows3d}}.
}
\examples{
# none yet
}
\author{
January Weiner, borrowed from from the \pkg{pca3d} package
}
\seealso{
\code{\link{arrows3d}}
}

