% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eqn.R
\name{Eqn}
\alias{Eqn}
\alias{Eqn_newline}
\title{Create a LaTeX Equation Wrapper}
\usage{
Eqn(..., number = TRUE, label = NULL, align = FALSE)

Eqn_newline()
}
\arguments{
\item{...}{expressions using \code{\link{cat}} and
return a \code{NULL} object to be wrapped in a LaTeX math
environment. Supplying a character vector will automatically wrap the
expression inside a call to \code{\link{cat}}}

\item{number}{logical; include equation number? Default: {TRUE}}

\item{label}{character vector specifying the LaTeX label to use (e.g., \code{eqn:myeqn})}

\item{align}{logical; use the \code{align} environment with explicit \code{&}. Default: {FALSE}}
}
\description{
This function is designed to produce LaTeX expressions that can be copied/pasted into documents or
used directly in \code{.Rmd} or \code{.qmd} documents to compile to equations.
It wraps the equations in either a \code{\\begin{equation} ...\\end{equation}} or
\code{\\begin{align} ...\\end{align}} environment.

In a code chunk, use the chunk options \code{results='asis', echo=FALSE}.
}
\examples{

Eqn('e=mc^2')
Eqn(cat('e=mc^2 \n')) # equivalent, but unnecessary

# Equation numbers & labels
Eqn('e=mc^2', number = FALSE)
Eqn('e=mc^2', label = 'eqn:einstein')
Eqn("X=U \\\\lambda V", label='eqn:svd')

# Multiple expressions
Eqn("e=mc^2", "X=U \\\\lambda V", label='eqn:svd')

# expressions that use cat() within their calls
Eqn(symbolicMatrix("u", "n", "k", lhs = 'SVD'),
    symbolicMatrix("\\\\lambda", "k", "k", diag=TRUE),
    symbolicMatrix("v", "k", "p", transpose = TRUE),
    label='eqn:svd')

# align equations using & operator
Eqn("X &= U \\\\lambda V",
    Eqn_newline(),
    symbolicMatrix("u", "n", "k", lhs = '&'),
    symbolicMatrix("\\\\lambda", "k", "k", diag=TRUE),
    symbolicMatrix("v", "k", "p", transpose = TRUE),
    align=TRUE)

#  matrix2latex() example
A <- matrix(c(2, 1, -1,
              -3, -1, 2,
              -2,  1, 2), 3, 3, byrow=TRUE)
b <- c(8, -11, -3)

matrix2latex(cbind(A,b)) |> Eqn()

# with showEqn()
showEqn(A, b, latex=TRUE) |> Eqn()

}
\seealso{
\code{\link{symbolicMatrix}}, \code{\link{matrix2latex}}
}
\author{
Phil Chalmers
}
