% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_under.R
\name{Eqn_helpers}
\alias{Eqn_helpers}
\alias{overset}
\alias{underset}
\alias{overbrace}
\alias{underbrace}
\alias{Eqn_overset}
\alias{Eqn_underset}
\alias{Eqn_overbrace}
\alias{Eqn_underbrace}
\title{Functions to Add Labels and/or Braces}
\usage{
overset(
  x,
  label,
  label.style = c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")
)

underset(
  x,
  label,
  label.style = c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")
)

overbrace(
  x,
  label = NULL,
  label.style = c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")
)

underbrace(
  x,
  label = NULL,
  label.style = c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")
)

Eqn_overset(
  x,
  label,
  label.style = c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")
)

Eqn_underset(
  x,
  label,
  label.style = c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")
)

Eqn_overbrace(
  x,
  label = NULL,
  label.style = c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")
)

Eqn_underbrace(
  x,
  label = NULL,
  label.style = c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")
)
}
\arguments{
\item{x}{a numeric or character matrix, or a character string LaTeX expression or 
a \code{"latexMatrix"} object}

\item{label}{a character string used as the label above or below the object \code{x}.
If missing, and a \code{"matrix"} object was passed, it's name is used as the label.
In LaTeX, these are rendered in a size appropriate for superscripts and subscripts,
but you can use a size modifier to change this, for example \code{'\\Large{"A"}'}.}

\item{label.style}{The name of a math font used to to typeset the label. One of
\code{c("mathbf", "mathrm", "mathit", "mathsf", "mathcal", "mathtt", " ")}.
The default, \code{"mathbf"} wraps the label inside \code{"\\mathbf{ }"}
commonly used for the name of a matrix.}
}
\value{
Returns a character vector containing the LaTeX expressions for the given operation. You can pass
       this to \code{\link{cat}} to display the result on the console, or include it inside a call
       to \code{Eqn} to typeset it.
}
\description{
These functions extend \code{\link{Eqn}} to add labels or braces over or under a LaTeX expression
or a \code{"latexMatrix"} object.
\itemize{
   \item{\code{overset} and \code{underset} typesets a label over or under an object}
   \item{\code{overbrace} and \code{underbrace} typesets a brace, with an optional label over or under an object}
}
Each of these have \code{Eqn_*} aliases to fit with the style of \code{\link{Eqn}} helper functions.

For example, given the matrix \code{A = matrix(1:4), 2, 2)}, the call \code{Eqn(overset(A, "A"))}
generates:
\preformatted{
\overset{\mathbf{A}} 
 { \begin{pmatrix} 
  1 & 3 \\ 
  2 & 4 \\ 
  \end{pmatrix}
 }
}

 When rendered in LaTeX, this produces:
 \deqn{
 \overset{\mathbf{A}} 
 { \begin{pmatrix} 
  1 & 3 \\ 
  2 & 4 \\ 
  \end{pmatrix}
 }
 }
 
 You can also use these for straight LaTeX expressions, such this equation showing and labeling
 the Hat matrix in regression. See the examples for the call to \code{underbrace} for this.
 \deqn{\mathbf{\hat{y}} =
       \underbrace{\mathbf{X}(\mathbf{X}^{\top}\mathbf{X})^{-1}
       \mathbf{X}^{\top}}_{\mathbf{\mathbf{H}}}\mathbf{y}}
}
\examples{
library(matlib)
A <- matrix(1:4, 2, 2)
B <- matrix(4:1, 2, 2)
AB <- A + B
Eqn(overset(A, "A"))
  #  missing label: uses the name of the object
Eqn(overset(A))

# test just a character LaTeX expression
Eqn('a', overset('=', '?'), 'b')

# a labelled latexMatrix equation
Eqn(overset(A, "A"), "+",
    overset(B, "B"), "=",
    underset(AB, "A+B"))
    
 # using a LaTeX expression as the label  
 Lambda <- latexMatrix("\\\\lambda", nrow=2, ncol=2,
                       diag=TRUE)
 Eqn(overset(Lambda, "\\\\Lambda"))

 # generate LaTeX expression for the Hat matrix, label as "H"
H <- "\\\\mathbf{X} (\\\\mathbf{X}^{\\\\top}\\\\mathbf{X})^{-1} \\\\mathbf{X}^{\\\\top}"
Eqn("\\\\mathbf{\\\\hat{y}} =", underbrace(H, "\\\\mathbf{H}"), "\\\\mathbf{y}")

# Combine this with overbrace
Eqn(overbrace(underbrace(H, "\\\\mathbf{H}"), "\\\\LARGE\\\\mathbf{\\\\hat{y}}"))

}
