% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{getYmult}
\alias{getYmult}
\title{Correct for aspect and coordinate ratio}
\usage{
getYmult()
}
\value{
The correction factor for the Y dimension.
}
\description{
Calculate a multiplication factor for the Y dimension to correct for unequal plot aspect and
coordinate ratios on the current graphics device.
}
\details{
\code{getYmult} retrieves the plot aspect ratio and the coordinate ratio for the current graphics device, calculates a
multiplicative factor to equalize the X and Y dimensions of a plotted graphic object.
}
\author{
Jim Lemon
}
