% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GramSchmidt.R
\name{GramSchmidt}
\alias{GramSchmidt}
\title{Gram-Schmidt Orthogonalization of a Matrix}
\usage{
GramSchmidt(X, normalize = TRUE)
}
\arguments{
\item{X}{a matrix}

\item{normalize}{logical; should the resulting columns be normalized to unit length?}
}
\value{
A matrix of the same size as \code{X}, with orthogonal columns
}
\description{
Carries out simple Gram-Schmidt orthogonalization of a matrix.
Treating the columns of the matrix \code{X} in the given order,
each successive column after the first is made orthogonal to all
previous columns by subtracting their projections on the current
column.
}
\examples{
(xx <- matrix(c( 1:3, 3:1, 1, 0, -1), 3, 3))
crossprod(xx)
(yy <- GramSchmidt(xx))
zapsmall(crossprod(yy))
}
\author{
Phil Chalmers
}

