\name{matlib-package}
\alias{matlib-package}
\alias{matlib}
\docType{package}
\title{
\packageTitle{matlib}
}
\description{
\packageDescription{matlib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{matlib}
\packageIndices{matlib}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{matlib}

Maintainer: \packageMaintainer{matlib}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
