% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regvec3d.R
\name{plot.regvec3d}
\alias{plot.regvec3d}
\alias{print.regvec3d}
\alias{summary.regvec3d}
\title{Plot method for regvec3d objects}
\usage{
\method{plot}{regvec3d}(x, y, dimension = c("3", "2"), col = c("black",
  "red", "blue", "magenta"), col.plane = "gray", show.plane = 2,
  show.marginal = FALSE, grid = FALSE, add = FALSE, ...)

\method{summary}{regvec3d}(object, ...)

\method{print}{regvec3d}(x, ...)
}
\arguments{
\item{x}{A \dQuote{regvec3d} object}

\item{y}{Ignored; only included for compatibility with the S3 generic}

\item{dimension}{Number of dimensions to plot: \code{"3"} (default) or \code{"2"}}

\item{col}{A vector of 4 colors}

\item{col.plane}{Color of the base plane in a 3D plot or axes in a 2D plot}

\item{show.plane}{If \code{show.plane > 0}, draws the base plane in a 3D plot; if \code{show.plane > 1},
the plane is drawn thicker}

\item{show.marginal}{If \code{TRUE} also draws lines showing the marginal relations of \code{y} on \code{x1} and on \code{x2}}

\item{grid}{If \code{TRUE}, draws a light grid on the base plane}

\item{add}{If \code{TRUE}, add to the current plot; otherwise start a new rgl or plot window}

\item{...}{Parameters passed down to functions [unused now]}

\item{object}{A \code{regvec3d} object for the \code{summary} method}
}
\value{
None

None
}
\description{
The plot method for \code{regvec3d} objects uses the low-level graphics tools in this package to draw 3D and 2D
vector diagrams reflecting the partial and marginal
relations of \code{y} to \code{x1} and \code{x2} in a bivariate multiple linear regression model,
\code{lm(y ~ x1 + x2)}.

The \code{summary} method prints the vectors and their vector lengths, followed by the \code{summary}
for the model.

Print method for regvec3d objects
}
\details{
A 3D diagram shows the vector \code{y} and the plane formed by the predictors,
\code{x1} and \code{x2}, where all variables are represented in deviation form, so that
the intercept need not be included.

A 2D diagram, using the first two columns of the result, can be used to show the projection
of the space in the \code{x1}, \code{x2} plane.
}
\examples{
if (require(car)) {
data("Duncan", package="car")
plot(regvec3d(prestige ~ income + education, data=Duncan))
}
}
\references{
Fox, J. (2016). \emph{Applied Regression Analysis and Generalized Linear Models}, Sage, Chapter 10.
}
\seealso{
\code{\link{regvec3d}}

Other vector diagrams: \code{\link{Proj}},
  \code{\link{arc}}, \code{\link{arrows3d}},
  \code{\link{corner}}, \code{\link{point_on_line}},
  \code{\link{regvec3d}}, \code{\link{vectors3d}},
  \code{\link{vectors}}
}

