% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotEqn.R
\name{plotEqn}
\alias{plotEqn}
\title{Plot Linear Equations
Shows what matrices \eqn{A, b} look like as the system of linear equations, \eqn{A x = b},
by plotting a line for each equation.}
\usage{
plotEqn(A, b, vars, xlim = c(-4, 4), ylim, col = 1:nrow(A), lwd = 2,
  lty = 1, axes = TRUE, labels = TRUE)
}
\arguments{
\item{A}{either the matrix of coefficients of a system of linear equations, or the matrix \code{cbind(A,b)}}

\item{b}{if supplied, the vector of constants on the right hand side of the equations}

\item{vars}{a numeric or character vector of names of the variables.
If supplied, the length must be equal to the number of unknowns in the equations.
The default is \code{paste0("x", 1:ncol(A)}.}

\item{xlim}{horizontal axis limits for the first variable}

\item{ylim}{horizontal axis limits for the second variable}

\item{col}{scalar or vector of colors for the lines, recycled as necessary}

\item{lwd}{scalar or vector of line widths for the lines, recycled as necessary}

\item{lty}{scalar or vector of line types for the lines, recycled as necessary}

\item{axes}{logical; draw horizontal and vertical axes through (0,0)?}

\item{labels}{logical, or a vector of character labels for the equations; not yet implemented}
}
\description{
Plot Linear Equations
Shows what matrices \eqn{A, b} look like as the system of linear equations, \eqn{A x = b},
by plotting a line for each equation.
}
\examples{
# consistent equations
A<- matrix(c(1,2,3, -1, 2, 1),3,2)
b <- c(2,1,3)
showeqn(A, b)
plotEqn(A,b)

# inconsistent equations
b <- c(2,1,6)
showeqn(A, b)
plotEqn(A,b)
}

