% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{printMatrix}
\alias{printMatrix}
\title{(Deprecated) Print a matrix, allowing fractions or LaTeX output}
\usage{
printMatrix(
  A,
  parent = TRUE,
  fractions = FALSE,
  latex = FALSE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{A}{A numeric matrix}

\item{parent}{flag used to search in the parent envir for suitable definitions of other arguments.
Set to \code{TRUE} (the default) if you want to only use the inputs provided.}

\item{fractions}{If \code{TRUE}, print numbers as rational fractions, using the \code{\link[MASS]{fractions}}
function; if you require greater accuracy, you can set the \code{cycles} (default 10)
and/or \code{max.denominator} (default 2000) arguments to \code{fractions} as a global option, e.g.,
\code{options(fractions=list(cycles=100, max.denominator=10^4))}.}

\item{latex}{If \code{TRUE}, print the matrix in LaTeX format}

\item{tol}{Tolerance for rounding small numbers to 0}
}
\value{
The formatted matrix
}
\description{
(Deprecated) Print a matrix, allowing fractions or LaTeX output
}
\examples{
A <- matrix(1:12, 3, 4) / 6
printMatrix(A, fractions=TRUE)
printMatrix(A, latex=TRUE)
}
\seealso{
\code{\link[MASS]{fractions}}
}
