% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{Proj}
\alias{Proj}
\title{Projection of Vector y on columns of X}
\usage{
Proj(y, X, list = FALSE)
}
\arguments{
\item{y}{a vector, treated as a one-column matrix}

\item{X}{a vector or matrix.  Number of rows of \code{y} and \code{X} must match}

\item{list}{logical; if FALSE, return just the projected vector; otherwise returns a list}
}
\value{
the projection of \code{y} on \code{X} (if \code{list=FALSE}) or a list with elements \code{y} and \code{P}
}
\description{
Fitting a linear model, \code{lm(y ~ X)}, by least squares can be thought of geometrically as the orthogonal projection of
\code{y} on the column space of \code{X}. This function is designed to allow exploration of projections
and orthogonality.
}
\details{
The projection is defined as \eqn{P y} where \eqn{P = X (X'X)^- X'}
and \eqn{X^-} is a generalized inverse.
}
\examples{
X <- matrix( c(1, 1, 1, 1, 1, -1, 1, -1), 4,2, byrow=TRUE)
y <- 1:4
Proj(y, X[,1])  # project y on unit vector
Proj(y, X[,2])
Proj(y, X)

# project unit vector on line between two points
y <- c(1,1)
p1 <- c(0,0)
p2 <- c(1,0)
Proj(y, cbind(p1, p2))

# orthogonal complements
y <- 1:4
yp <-Proj(y, X, list=TRUE)
yp$y
P <- yp$P
IP <- diag(4) - P
yc <- c(IP \%*\% y)
crossprod(yp$y, yc)

# P is idempotent:  P P = P
P \%*\% P
all.equal(P, P \%*\% P)
}
\seealso{
Other vector diagrams: 
\code{\link{arc}()},
\code{\link{arrows3d}()},
\code{\link{circle3d}()},
\code{\link{corner}()},
\code{\link{plot.regvec3d}()},
\code{\link{pointOnLine}()},
\code{\link{regvec3d}()},
\code{\link{vectors}()},
\code{\link{vectors3d}()}
}
\author{
Michael Friendly
}
\concept{vector diagrams}
