% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matlib.R
\name{R}
\alias{R}
\title{Rank of a Matrix}
\usage{
R(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
rank of \code{X}
}
\description{
Returns the rank of a matrix \code{X}, using the QR decomposition, \code{qr()}.
Included here as a simple function, because \code{rank} does something different
and it is not obvious what to use for matrix rank.
}

