% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{vectors}
\alias{vectors}
\title{Title Draw Geometric Vectors}
\usage{
vectors(X, origin = c(0, 0), lwd = 2, angle = 10, length = 0.15,
  labels = TRUE, cex.lab = 1.5, pos.lab = 4, ...)
}
\arguments{
\item{X}{a vector or two-column matrix representing a set of geometric vectors; if a matrix, one vector is drawn for each row}

\item{origin}{the origin from which they are drawn, a vector of length 2.}

\item{lwd}{line width(s) for the vectors, a constant or vector of length equal to the number of rows of \code{X}.}

\item{angle}{the \code{angle} argument passed to \code{\link[graphics]{arrows}} determining the angle of arrow heads.}

\item{length}{the \code{length} argument passed to \code{\link[graphics]{arrows}} determining the length of arrow heads.}

\item{labels}{a logical or a character vector of labels for the vectors. If \code{TRUE} and \code{X} is a matrix,
labels are taken from \code{rownames(X)}. If \code{NULL}, no labels are drawn.}

\item{cex.lab}{character expansion applied to vector labels}

\item{pos.lab}{label position relative to the label point}

\item{...}{other arguments passed on to graphics functions}
}
\value{
none
}
\description{
This function draws vectors in a 2D plot, in a way that facilitates constructing vector diagrams.
}
\examples{
# shows addition of vectors
u <- c(3,1)
v <- c(1,3)
sum <- u+v

xlim <- c(0,5)
ylim <- c(0,5)
# proper geometry requires asp=1
plot( xlim, ylim, type="n", xlab="X", ylab="Y", asp=1)
abline(v=0, h=0, col="gray")

vectors(rbind(u,v,`u+v`=sum), col=c("red", "blue", "purple"))
# show the opposing sides of the parallelogram
vectors(sum, origin=u, col="red", lty=2)
vectors(sum, origin=v, col="blue", lty=2)

# projection of vectors
vectors(Proj(v,u), labels="P(v,u)", lwd=3)
vectors(v, origin=Proj(v,u))
}

